fixed data types

+++ src/DataTypes.java
@@ -2,10 +2,9 @@ import java.util.List;
 
 public class DataTypes {
     public static long sum(List<Integer> numbers) {
-
-        int s = 0;
+        long s = 0;
         // below is a "foreach" loop which iterates through numbers
-        for (int x : numbers) {
+        for (long x : numbers) {
             s += x;
         }
         return s;
